requirements.txt
asgiref==3.7.2
Brotli==1.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-cron==0.6.0
django-filter==23.3
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
Markdown==3.5
mysqlclient==2.2.0
packaging==23.2
Pygments==2.16.1
PyJWT==2.8.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.6.0
